resolve complex architecture - Remove complex code standard - Follow KISS principle
