--- .vscode/settings.json
@@ -1,13 +1,0 @@
-{
-  "cSpell.words": [
-    "antd",
-    "hookform",
-    "LEMONSQUEEZY",
-    "NEXTAUTH",
-    "signin",
-    "Signout",
-    "Signup",
-    "Squeezy",
-    "USEPLUNK"
-  ]
-}
