+++ src/app/(landing-page)/(pages)/layout.tsx
