+++ src/app/(landing-page)/(pages)/signin/page.tsx
@@ -1,7 +1,7 @@
 import { redirect } from "next/navigation";
 
 import { auth } from "@/auth";
-import { SignUpForm } from "./form";
+import { SignInForm } from "./form";
 
 export default async function SignInPage() {
   const session = await auth();
@@ -11,8 +11,8 @@ export default async function SignInPage() {
   return (
     <main className="py-10 container mx-auto">
       <div className="mx-auto w-full md:max-w-xl">
-        <h3 className="font-bold text-xl mb-3">Signup Page</h3>
-        <SignUpForm />
+        <h3 className="font-bold text-xl mb-3">Sign in</h3>
+        <SignInForm />
       </div>
     </main>
   );
