+++ src/app/(landing-page)/(pages)/signup/form.tsx
@@ -6,7 +6,6 @@ import { useState } from "react";
 import { InfoIcon } from "lucide-react";
 import { useForm } from "react-hook-form";
 import { useRouter } from "next/navigation";
-import { revalidatePath } from "next/cache";
 import { zodResolver } from "@hookform/resolvers/zod";
 
 import { Button } from "@/components/ui/button";
@@ -19,23 +18,25 @@ import {
   FormLabel,
   FormMessage,
 } from "@/components/ui/form";
-import { signIn } from "@/auth/helpers";
 import { Input } from "@/components/ui/input";
 import { toast } from "@/components/ui/use-toast";
+import { signUp } from "@/auth/actions";
 
 const formSchema = z.object({
+  name: z.string().trim().min(4, "Name is required"),
   email: z.string().trim().email(),
   password: z.string().min(4, "Password is minimum 4 characters."),
 });
 
-export const SigninForm = () => {
+export const SignUpForm = () => {
   const router = useRouter();
 
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
+  const [isSignedUp, setIsSignedUp] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
     defaultValues: {
+      name: "",
       email: "",
       password: "",
     },
@@ -44,41 +45,36 @@ export const SigninForm = () => {
   const onSubmit = async (values: z.infer<typeof formSchema>) => {
     setIsLoading(true);
 
-    try {
-      const response = await signIn(values);
+    const response = await signUp(values);
 
-      if (response?.success) {
-        router.push("/dashboard");
-        toast({
-          title: "Sent",
-          description: response.message,
-        });
-        form.reset();
-        setIsLoggedIn(true);
-        // revalidatePath("/");
-      } else {
-        toast({
-          variant: "destructive",
-          title: "Error",
-          description: response?.message,
-        });
-      }
-    } catch (error) {
-      // console.log(error);
-    } finally {
-      setIsLoading(false);
+    if (response?.success) {
+      router.push("/dashboard");
+      toast({
+        title: "Sent",
+        description: response.message,
+      });
+      form.reset();
+      setIsSignedUp(true);
+    } else {
+      toast({
+        variant: "destructive",
+        title: "Error",
+        description: response?.message,
+      });
     }
+
+    setIsLoading(false);
   };
 
-  return isLoggedIn ? (
+  return isSignedUp ? (
     <div
       className="flex items-center p-4 mb-4 text-sm text-green-800 rounded-lg bg-green-50 dark:bg-gray-800 dark:text-green-400 gap-1 mt-5"
       role="alert"
     >
       <InfoIcon className="w-4 h-4" />
       <span className="sr-only">Info</span>
-      <span className="font-medium">Success alert!</span> You have been logged
-      in successfully.
+      <span className="font-medium">Success alert!</span> You have been signed
+      up successfully.
       <Link href="/dashboard" className="underline">
         Visit dashboard
       </Link>
@@ -88,6 +84,20 @@ export const SigninForm = () => {
       <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
         <FormField
           control={form.control}
+          name="name"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Name</FormLabel>
+              <FormControl>
+                <Input type="text" placeholder="john" {...field} />
+              </FormControl>
+              <FormDescription>Enter your name.</FormDescription>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <FormField
+          control={form.control}
           name="email"
           render={({ field }) => (
             <FormItem>
@@ -95,7 +105,7 @@ export const SigninForm = () => {
               <FormControl>
                 <Input type="email" placeholder="john@email.com" {...field} />
               </FormControl>
-              <FormDescription>Enter your email to login.</FormDescription>
+              <FormDescription>Enter your email.</FormDescription>
               <FormMessage />
             </FormItem>
           )}
@@ -117,7 +127,7 @@ export const SigninForm = () => {
           )}
         />
         <Button type="submit" disabled={isLoading}>
-          Signin
+          Sign up
         </Button>
       </form>
     </Form>
