+++ src/app/(landing-page)/_components/footer/footer.tsx
