+++ src/auth/routes.ts
@@ -1,0 +1,6 @@
+export const SIGNIN = "/signin";
+export const ROOT = "/";
+
+export const PUBLIC_ROUTES = ["/signin", "/signup"];
+
+export const PROTECTED_SUB_ROUTES = ["/dashboard"];
