--- src/lib/routes/index.ts
@@ -1,6 +1,0 @@
-export const LOGIN = '/signin';
-export const ROOT = '/';
-
-export const PUBLIC_ROUTES = ['/signin', '/signup'];
-
-export const PROTECTED_SUB_ROUTES = ['/dashboard'];
