--- src/lib/token/generate-token.ts
@@ -1,5 +1,0 @@
-import jwt from "jsonwebtoken";
-
-export const generateToken = () => {
-  return jwt.sign({}, process.env.INVITATION_TOKEN_SECRET);
-};
