--- src/shared/hooks/index.ts
@@ -1,1 +1,0 @@
-// reusable hooks goes here
