--- src/shared/infrastructure/StandardReturn.ts
@@ -1,21 +1,0 @@
-export type StandardReturn = {
-  status: 'success' | 'error' | 'promise' | 'partial';
-  errors: string[]; //managed by code
-  fails?: string[]; //to display to user
-};
-
-export type StandardQueryReturn<T> = StandardReturn & {
-  data: T | null;
-};
-
-export type StandardCreateReturn = StandardReturn & {
-  data: string;
-};
-
-export type StandardMutationReturn = StandardReturn & {
-  data: string | boolean;
-};
-
-export type StandardBulkDeleteReturn = StandardReturn & {
-  data: string[];
-};
